Add source attribute
Before this change the motd::fragment defined type only had a content
attribute and didn't allow to specify a file.

This change modifies the defined type to allow using a source or content
attribute, similar to how the file type works.

The defined type will raise an error if neither or both attributes are
provided. In other words, it requires either content or source.